+++ src/Devices/ElectricalSource/Amplifier/SimulatedAmplifier.jl
@@ -3,7 +3,7 @@ export SimulatedAmplifier, SimulatedAmplifierParams
 Base.@kwdef struct SimulatedAmplifierParams <: DeviceParams
   channelID::String
   mode::AmplifierMode = AMP_VOLTAGE_MODE # This should be the safe default
-	voltageMode::AmplifierVoltageMode = AMP_LOW_VOLTAGE_MODE # This should be the safe default
+	powerSupplyMode::AmplifierPowerSupplyMode = AMP_LOW_POWER_SUPPLY # This should be the safe default
 	matchingNetwork::Integer = 1
 end
 SimulatedAmplifierParams(dict::Dict) = params_from_dict(SimulatedAmplifierParams, dict)
@@ -13,14 +13,14 @@ Base.@kwdef mutable struct SimulatedAmplifier <: Amplifier
 
   state::Bool = false
   mode::AmplifierMode = AMP_VOLTAGE_MODE
-  voltageMode::AmplifierVoltageMode = AMP_LOW_VOLTAGE_MODE
+  powerSupplyMode::AmplifierPowerSupplyMode = AMP_LOW_POWER_SUPPLY
   network::Integer = 1
 end
 
 function _init(amp::SimulatedAmplifier)
   # Set values given by configuration
 	mode(amp, amp.params.mode)
-	voltageMode(amp, amp.params.voltageMode)
+	powerSupplyMode(amp, amp.params.powerSupplyMode)
 	matchingNetwork(amp, amp.params.matchingNetwork)
 end
 
@@ -33,8 +33,8 @@ turnOn(amp::SimulatedAmplifier) = amp.state = true
 turnOff(amp::SimulatedAmplifier) = amp.state = false
 mode(amp::SimulatedAmplifier) = amp.mode
 mode(amp::SimulatedAmplifier, mode::AmplifierMode) = amp.mode = mode
-voltageMode(amp::SimulatedAmplifier) = amp.voltageMode
-voltageMode(amp::SimulatedAmplifier, mode::AmplifierVoltageMode) = amp.voltageMode = mode
+powerSupplyMode(amp::SimulatedAmplifier) = amp.powerSupplyMode
+powerSupplyMode(amp::SimulatedAmplifier, mode::AmplifierPowerSupplyMode) = amp.powerSupplyMode = mode
 matchingNetwork(amp::SimulatedAmplifier) = amp.network
 matchingNetwork(amp::SimulatedAmplifier, network::Integer) = amp.network = network
 temperature(amp::SimulatedAmplifier) = 25.0u"°C"